+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameValidationService.java
@@ -580,9 +580,6 @@ public class GameValidationService implements InitializingBean, PreferenceChange
         switch (altColorType) {
             case pal: {
                 if (isValidationEnabled(game, CODE_ALT_COLOR_FILES_MISSING)) {
-//          if (altColor.contains("pin2dmd.pal") && !altColor.contains("pin2dmd.vni")) {
-//            result.add(GameValidationStateFactory.create(CODE_ALT_COLOR_FILES_MISSING, "pin2dmd.vni"));
-//          }
                     if (!altColor.contains("pin2dmd.pal") && altColor.contains("pin2dmd.vni")) {
                         result.add(ValidationStateFactory.create(CODE_ALT_COLOR_FILES_MISSING, "pin2dmd.pal"));
                     }
@@ -601,7 +598,7 @@ public class GameValidationService implements InitializingBean, PreferenceChange
                 break;
             }
             case cROMc: {
-                String name = (!StringUtils.isEmpty(game.getRomAlias()) ? game.getRomAlias() : game.getRom()) + "." + UploaderAnalysis.CROMC_SUFFIX;
+                String name = game.getRom() + "." + UploaderAnalysis.CROMC_SUFFIX;
                 if (game.isZenGame()) {
                     String gameNameForAltSound = dofLinxService.getGameNameForAltColor(game);
                     name = gameNameForAltSound + "." + UploaderAnalysis.CROMC_SUFFIX;